+++ domain/src/main/java/fr/insee/rem/domain/ports/spi/PartitionSurveyUnitLinkPersistencePort.java
@@ -21,4 +21,6 @@ public interface PartitionSurveyUnitLinkPersistencePort {
     List<SuIdMappingRecord> findSurveyUnitIdsMappingTableByPartitionId(Long partitionId);
 
     long countByPartitionId(Long partitionId);
+
+    void removeSurveyUnitsFromPartition(Long partitionId);
 }