+++ domain/src/test/java/fr/insee/rem/domain/service/PartitionServiceTest.java
@@ -1,10 +1,15 @@
 package fr.insee.rem.domain.service;
 
+import fr.insee.rem.domain.dtos.Context;
 import fr.insee.rem.domain.dtos.PartitionDto;
+import fr.insee.rem.domain.dtos.PartitionSurveyUnitLinkDto;
+import fr.insee.rem.domain.dtos.SurveyUnitDto;
 import fr.insee.rem.domain.exception.PartitionAlreadyExistsException;
 import fr.insee.rem.domain.exception.PartitionNotFoundException;
 import fr.insee.rem.domain.ports.api.PartitionServicePort;
 import fr.insee.rem.domain.service.mock.PartitionPersistenceInMemory;
+import fr.insee.rem.domain.service.mock.PartitionSurveyUnitLinkPersistenceInMemory;
+import fr.insee.rem.domain.service.mock.SurveyUnitPersistenceInMemory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +19,11 @@ import java.util.List;
 class PartitionServiceTest {
 
     PartitionPersistenceInMemory partitionPersistenceInMemory = new PartitionPersistenceInMemory();
-    PartitionServicePort partitionService = new PartitionServiceImpl(partitionPersistenceInMemory);
+    SurveyUnitPersistenceInMemory surveyUnitPersistenceInMemory = new SurveyUnitPersistenceInMemory();
+    PartitionSurveyUnitLinkPersistenceInMemory partitionSurveyUnitLinkPersistenceInMemory =
+            new PartitionSurveyUnitLinkPersistenceInMemory(partitionPersistenceInMemory, surveyUnitPersistenceInMemory);
+    PartitionServicePort partitionService = new PartitionServiceImpl(partitionPersistenceInMemory,
+            partitionSurveyUnitLinkPersistenceInMemory);
 
     @Test
     void shouldCreatePartitionFromLabel() {
@@ -109,4 +118,38 @@ class PartitionServiceTest {
         Assertions.assertEquals(partitionOne, allPartitions.get(0));
         Assertions.assertEquals(partitionTwo, allPartitions.get(1));
     }
+
+    @Test
+    void shouldReturnPartitionNotFoundExceptionWhenDeleteSurveyUnitsFromNotExistingPartition() {
+        // Given
+        Long notExistingPartitionId = 99L;
+
+        // When + Then
+        PartitionNotFoundException exception = Assertions.assertThrows(PartitionNotFoundException.class,
+                () -> partitionService.emptyPartitionById(notExistingPartitionId));
+        Assertions.assertEquals(String.format("Partition [%s] doesn't exist", notExistingPartitionId),
+                exception.getMessage());
+    }
+
+    @Test
+    void shouldDeleteSurveyUnitsFromPartition() {
+        // Given
+        PartitionDto existingPartition = partitionService.createPartition("existing partition");
+        List<SurveyUnitDto> surveyUnitsToImport =
+                List.of(SurveyUnitDto.builder().context(Context.HOUSEHOLD).externalId("00001").build(),
+                        SurveyUnitDto.builder().context(Context.HOUSEHOLD).externalId("00002").build());
+        List<PartitionSurveyUnitLinkDto> importedPartitionSurveyUnits =
+                partitionSurveyUnitLinkPersistenceInMemory.saveAll(existingPartition.getPartitionId(),
+                        surveyUnitsToImport);
+
+        // When
+        partitionService.emptyPartitionById(existingPartition.getPartitionId());
+
+        // Then
+        List<PartitionSurveyUnitLinkDto> links =
+                partitionSurveyUnitLinkPersistenceInMemory.findSurveyUnitsByPartitionId(existingPartition.getPartitionId());
+        Assertions.assertEquals(0, links.size());
+    }
+
+
 }